Cuarta versión de las gráficas de los cuestionarios: layout de los botones dependiendo del ancho de la ventana y leyenda oculta cuando hay poco espacio vertical. Uso de los producer de vico para mejorar el rendimiento. Disgregado el flujo de estados y de operaciones en el viewmodel.
Implementada la función por el cual los desplegables tienen como valor por defecto el que trae el viewmodel (evitando fallos al girar la pantalla). Recuperadas las previews e incluidas nuevas.

Como efecto colateral tener en cuenta el ancho y el alto de la ventana y que este puede cambiar al girar el dispositivo, se ha adaptado todo el diseño responsive a la api de Window Size de Material Design 3 y se ha eliminado que sea una informacion singleton

+++ app/src/main/java/es/upm/bienestaremocional/app/domain/processing/Questionnaire.kt
@@ -1,35 +1,34 @@
 package es.upm.bienestaremocional.app.domain.processing
 
 import es.upm.bienestaremocional.app.data.database.entity.QuestionnaireEntity
-import java.time.Instant
 
 /**
- * Transform a list of [QuestionnaireEntity] into another of type [Pair]<[Instant],[Int]>
- * Int data represent is the score of these day, computed as the average of the scores of the day
+ * Transform a list of [QuestionnaireEntity] into another of type [Pair]<[Long],[Float]>
+ * Float data represent is the score of these day, computed as the average of the scores of the day
  * @param records list to transform
  * @return records transformed
  */
-fun aggregateEntriesPerDay(records: List<QuestionnaireEntity>): List<Pair<Long,Int>> =
+fun aggregateEntriesPerDay(records: List<QuestionnaireEntity>): List<Pair<Long,Float>> =
     aggregateEntriesPer(records) { truncateTimestampToDay(it) }
 
 /**
- * Transform a list of [QuestionnaireEntity] into another of type [Pair]<[Instant],[Int]>
- * Int data represent is the score of these week, computed as the average of the scores of the week
+ * Transform a list of [QuestionnaireEntity] into another of type [Pair]<[Long],[Float]>
+ * Float data represent is the score of these week, computed as the average of the scores of the week
  */
-fun aggregateEntriesPerWeek(records: List<QuestionnaireEntity>): List<Pair<Long,Int>> =
+fun aggregateEntriesPerWeek(records: List<QuestionnaireEntity>): List<Pair<Long,Float>> =
     aggregateEntriesPer(records) { truncateTimestampToWeek(it) }
 
 /**
- * Transform a list of [QuestionnaireEntity] into another of type [Pair]<[Instant],[Int]>
- * Int data represent is the score of these month, computed as the average of the scores of the month
+ * Transform a list of [QuestionnaireEntity] into another of type [Pair]<[Long],[Float]>
+ * Float data represent is the score of these month, computed as the average of the scores of the month
  */
-fun aggregateEntriesPerMonth(records: List<QuestionnaireEntity>): List<Pair<Long,Int>> =
+fun aggregateEntriesPerMonth(records: List<QuestionnaireEntity>): List<Pair<Long,Float>> =
     aggregateEntriesPer(records) { truncateTimestampToMonth(it) }
 
 private fun aggregateEntriesPer(records: List<QuestionnaireEntity>, truncate : (Long) -> Long):
-        List<Pair<Long,Int>>
+        List<Pair<Long,Float>>
 {
-    val result = mutableListOf<Pair<Long,Int>>()
+    val result = mutableListOf<Pair<Long,Float>>()
     val buffer = mutableMapOf<Long,MutableList<Int>>()
 
     records.forEach { record ->
@@ -44,7 +43,7 @@ private fun aggregateEntriesPer(records: List<QuestionnaireEntity>, truncate : (
     }
 
     buffer.forEach { entry ->
-        val score = entry.value.sum() / entry.value.size
+        val score = entry.value.sum().toFloat() / entry.value.size
         result.add(Pair(entry.key,score))
     }
